Cursors should be int64, not int

+++ friends_followers.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Cursor struct {
-	Previous_cursor     int
+	Previous_cursor     int64
 	Previous_cursor_str string
 
 	Ids []int64
